Fine-tune lrclib requests

+++ providers/lrclib/src/main/kotlin/app/vitune/providers/lrclib/models/Track.kt
@@ -9,11 +9,11 @@ data class Track(
     val id: Int,
     val trackName: String,
     val artistName: String,
-    val duration: Long,
+    val duration: Double,
     val plainLyrics: String?,
     val syncedLyrics: String?
 )
 
 internal fun List<Track>.bestMatchingFor(title: String, duration: Duration) =
-    firstOrNull { it.duration == duration.inWholeSeconds }
+    firstOrNull { it.duration.toLong() == duration.inWholeSeconds }
         ?: minByOrNull { abs(it.trackName.length - title.length) }
